Updating slice pose interpolation to no longer rely on matrix decomposition and composition
Ensuring splash system doesn't use a null pointer

+++ HoloIntervention/Content/Systems/Splash/SplashSystem.h
@@ -64,7 +64,7 @@ namespace HoloIntervention
 
       Windows::Foundation::Numerics::float3               m_position;
 
-      double                                              m_welcomeTimerSec = 0.f;
+      float                                               m_welcomeTimerSec = 0.f;
       std::wstring                                        m_splashImageFilename = L"Assets\\Images\\HoloIntervention.png";
 
       static const float                                  LERP_RATE;
